requirements.txt
# Oracle OPC P6 Data Pipeline - Dependencies

# HTTP & OAuth2
requests==2.31.0
requests-oauthlib==1.3.1
urllib3==2.1.0

# Oracle Database
python-oracledb==2.1.2

# MinIO / S3
minio==7.2.7

# Apache Airflow (core + providers)
apache-airflow==2.8.4
apache-airflow-providers-oracle==3.10.0
apache-airflow-providers-amazon==8.18.0
apache-airflow-providers-http==4.10.0

# Data processing
pandas==2.1.4
pyarrow==14.0.2
numpy==1.26.3

# Configuration
PyYAML==6.0.1
python-dotenv==1.0.0

# Retry / resilience
tenacity==8.2.3

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-mock==3.12.0
pytest-cov==4.1.0
responses==0.24.1

# Linting
flake8==7.0.0
black==23.12.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
